This is version 1.0 of autoInflate.py
This is the main script that will use the SleepQ library (from here on GITHUB) to connect to your SleepNumber bed using Sleep IQ API to force the bed to inflate on a timed interval to keep it from going flat overnight when the bed has a slow leak. You can tune the script yourself by changing the variables default values at the top of the script to meet your needs depending on how fast the leak is on your mattress.